+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Basic exceptions for git providers.
-- Clone repositories from a remote host.
+- Clone repositories from a remote host (From git providers).
 
 ## [0.0.3] 2024-01-28